+++ packages/components/modules/messages/common/graphql/fragments/AddContactToGroupItem.ts
@@ -1,0 +1,14 @@
+import { graphql } from 'react-relay'
+
+export const AddContactToGroupItemFragment = graphql`
+  fragment AddContactToGroupItemFragment on ChatRoom
+  @argumentDefinitions(contactProfileId: { type: "ID!" }) {
+    id
+    title
+    participantsCount
+    image(width: 100, height: 100) {
+      url
+    }
+    isParticipant(profileId: $contactProfileId)
+  }
+`