+++ packages/components/modules/messages/native/AddContactToGroupPage/GroupsList/GroupListItem/types.ts
@@ -1,0 +1,7 @@
+import { AddContactToGroupItemFragment$key } from '../../../../../../__generated__/AddContactToGroupItemFragment.graphql'
+
+export type GroupListItemProps = {
+  roomRef: AddContactToGroupItemFragment$key
+  selected: boolean
+  onToggle: (groupId: string) => void
+}