+++ packages/components/modules/messages/native/AddContactToGroupPage/GroupsList/styles.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  contentContainer: {
+    paddingBottom: 40,
+  },
+})