+++ packages/components/modules/messages/native/SingleChatDetailsPage/Options/type.ts
@@ -1,6 +1,7 @@
 export interface OptionsProps {
   handleArchiveChat: () => void
   handleGoToProfile: () => void
+  handleAddContactToGroup: () => void
   isArchiveMutationInFlight: boolean
   isArchived: boolean
 }