+++ packages/wagtail/CHANGELOG.md
@@ -1,12 +1,14 @@
 # @baseapp-frontend/wagtail
 
-## 1.0.63
+## 2.0.0
 
 ### Patch Changes
 
-- Updated dependencies [5200c84]
+- Updated dependencies [007b2ae]
 - Updated dependencies [c2f042d]
-  - @baseapp-frontend/design-system@1.4.1
+  - @baseapp-frontend/utils@4.2.0
+  - @baseapp-frontend/design-system@2.0.0
+  - @baseapp-frontend/graphql@2.0.0
 
 ## 1.0.62